reorg runtime structure
