+++ src/physicalai/runtime/action_sources/__init__.py
@@ -1,0 +1,12 @@
+# Copyright (C) 2026 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Action source implementations pluggable into RobotRuntime."""
+
+from __future__ import annotations
+
+from physicalai.runtime.action_sources.base import ActionSource
+from physicalai.runtime.action_sources.policy import PolicySource
+from physicalai.runtime.action_sources.teleop import TeleopSource
+
+__all__ = ["ActionSource", "PolicySource", "TeleopSource"]
