+++ src/physicalai/runtime/callbacks/__init__.py
@@ -1,0 +1,20 @@
+# Copyright (C) 2025-2026 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Shipped callback implementations for the runtime callback bus."""
+
+from __future__ import annotations
+
+from physicalai.runtime.callbacks.async_callback import AsyncCallback
+from physicalai.runtime.callbacks.console import ConsoleCallback
+from physicalai.runtime.callbacks.jsonl import JsonlCallback
+from physicalai.runtime.callbacks.low_pass import LowPassFilterCallback
+from physicalai.runtime.callbacks.rerun import RerunCallback
+
+__all__ = [
+    "AsyncCallback",
+    "ConsoleCallback",
+    "JsonlCallback",
+    "LowPassFilterCallback",
+    "RerunCallback",
+]
